+++ plugins/postcss-custom-properties/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS Custom Properties
 
+### Unreleased
+
+- Improve plugin performance
+
 ### 13.0.0 (November 14, 2022)
 
 - Updated: Support for Node v14+ (major).